coredns: Integrate external DNS with the existing coredns server
The internal DNS is not leaked: The external DNS is served on both
port 53 and port 5353. Only 5353 is open to the world.
This also means that the cluster can rely on the external DNS names
to always resolve since it solely hinges on the reachability of the
central coredns. No external glue-records and port forwards or router
configs needed.