Remove longhorn workload by relying on hostPath mounts
smallstep does not need db persistence for anything except revocation,
which doesn't really makes sense anyways because we have disabled renewal.

If apt-cacher-ng happens to be schedule on another host, let it. It's
not really a big deal to have the cache on more than one machine.

We force the image-registry to run on the control-plane. This will
be a problem once we support multiple masters